+++ app/Traits/Searchable.php
@@ -35,11 +35,13 @@ trait Searchable
 
         $builder->whereRaw("`$tablePrefix{$this->getTable()}`.`account_id` = $accountId AND ($queryString) $whereCondition");
         $builder->orderByRaw($orderBy);
+
         if ($sortOrder) {
             $builder->sortedBy($sortOrder);
         }
-        $builder->select(array_map(function ($column) use ($tablePrefix) {
-            return "$tablePrefix{$this->getTable()}.$column";
+
+        $builder->select(array_map(function ($column) {
+            return "{$this->getTable()}.$column";
         }, $this->return_from_search));
 
         return $builder->paginate($limitPerPage);